renaming the updated icons
